Move update effect, convert appStatsReducer to effect (WIP)
