+++ src/frontend/packages/store/src/helpers/store-helpers.ts
@@ -33,7 +33,6 @@ export function getFavoriteInfoObservable(store: Store<AppState>): Observable<IF
   );
 }
 
-// TODO: RC move to CF
 export const getRowMetadata = (entity: APIResource | MultiActionListEntity) => {
   if (entity instanceof MultiActionListEntity) {
     return entity.entity.metadata ? entity.entity.metadata.guid : null;
