+++ src/frontend/packages/store/src/types/api.types.ts
@@ -7,7 +7,6 @@ export interface EntityInfo<T = any> {
   entity: T;
 }
 
-// TODO: RC move to CF
 export interface APIResource<T = any> {
   metadata: APIResourceMetadata;
   entity: T;
